Many changes outlined below
The minor changes are just optimized imports, changing some class level
variables to local and more.

The first major change is that the data type for a JutsuGesture.allData
has been changed to an arraylist of arraylists of floats. This is done
to simplify many things but mainly it allows me to easily compare the
packets of data presented from onSensorChanged. As this data structure
was heavily used, making this change was major. To help with the fluency
of the change, on each sensor changed I immediately take the float[]
provided and convert it into an ArrayList<Float>.

With this new set of eventValues, I try and compare the previous data
that the sensor presented with the new data. This is so that tiny sensor
changed do not take up processing power by creating a new JutsuGesture.
Because of this change, the display now works and I can recognise
'major' wrist movements.

Added method to create a Jutsu Gesture and add it to activeGestures.

Created DataAnalyser class to contain methods that involve comparing the
standard data structures of this app: ArrayList<Float>.