Added new methods, added audio resources in /raw
One of the methods added was the init method for ProcessDataThread, the
other was a method in the DataAnalyser class that allows you to get the
difference between two sets of accelerometer data that has already been
converted.
Created and added audio clips in the form of resources under /raw
folder. All have the mp3 file extension.