Huge project wide changes, outlined below.
Once I came across the issue of creating a huge data structure to
represent a JutsuGesture, I found my current methodology just plain
stupid.

Previously, I was creating new JutsuGestures all the time (when major
data comes in from accelerometer) which was registered at all times. The
app was slow and laggy. I had to process almost 200 objects every time
the sensor changed and I didnt even start comparing these objects to
actual gesture objects. It only worked because I used an AsyncTask
thread to do most of the processing.

I took a step back, and tried to think of a better way. Enter my
notepad. Output a better way.

This commit is the first step(quite a big one) to transforming the
project into my solution.

The solution has multiple points.

First, I will be using 7 JSON files to store the JutsuGesture data, BUT
these will be emptied every time the app is opened. The interaction with
these files will probably be the heaviest processing required in the new
solution and it is so much lighter than the previous.

I will not go into too much detail here. Basically, the app only listens
for Jutsu Gestures once clicked and this enables the user to set what
classifies a Jutsu hand sign every time they reboot the app. Therefore,
the sensors are only listening for 5 seconds and only 1 justu gesture is
created in that time and analysed afterwards.