Work on JSON file interactions
Now there are 7 JSON files; one for each jutsu gesture.

Using the android resource system (so I never have to worry about one of
the filepaths) and the gson library (very lightweight and easy to use) I
am building methods in my JsonInteraction class that will be used by
JutsuGesture after the 5 seconds to do things such as clear, read and
write to the JSON file that belongs to that JutsuGesture.