Project wide changes outlined below;
First in ImageAdapter, I am not using the ProcessDataTask to call the
now-heavy method updateData.

As hinted at above, JutsuGesture had many changes. Smaller ones include
optimized imports and making certain variables private or local to their
usage. One of the major changes was that, as I created
defined_jutsu_data.json, I added into the switch statement a string
definition that would act as a key for the JsonInteraction class in
order to grab the correct data from the json file. The second big
change was the comletion (I hope) of updateData. What it now does is
heavily documented so you should be able to check it out and see what
was done with explainations.

Lastly in JutsuGesture, the playAudioClip method was completed and seems
to function.

The main changes made in JsonInteraction all revolved around the change
from 7 json files into one. Everything has been tested with this change
and seems to work.

Lastly, ProcessDataTask was cut down alot as its current only function
is to run updateData. This method is the heaviest task in the app as of
now and must be run in the AsyncTask.