Main app functionality complete!!
The jutsu data has been gathered via another app I built and put onto
the watch called Gather Jutsu Data. This allowed me to write sequences
of data to a file that I had access to of named gestures.

This data was added to defined_jutsu_data in this app, and I spent a
stupid amount of time tweaking the allowed number of errors and each
gestures threshold. I will admit it is still a bit janky but some of the
gestures are fairly consistent.

App icon was drawn, added to resources and finally made the official
icon of the app. The original image is linked in README.