Finished front end of app
This involved adding an intro screen and background music for the app.
The first addition meant that I had to increment my pager positions. The
background music plays at a lower volume than normal and lowers even
further while the jutsu gesture is listening to signify to the user that
it is listening.

I updated some variable names in this commit too. Lastly, if the user
presses the screen while on the intro screen, a Toast message pops up
letting the user know to swipe right.